hw: clean up nested page table switching
On x86_64, the current `Vm_page_table` implementation encapsulates the
nested page table implementation for the used virtualization type. This
pulled side concerns such as the caching of the used virtualization type
into the `Vm_page_table` implementation.
While such a runtime distinction is necessary on x86, it may better be
handled outside of the nested page table container.

To address this, first move virtualization type caching to
`Virtualization_support`, so that the type can be checked without
running the detection repeatedly.
Second, setup and call the paged table implementation needed for the
used virtualization type directly from the VM Session Component instead
of encapsulating the runtime selection in the x86 `Vm_page_table`.

Fixes #5221