hw: sepearate Vm_session_component for SVM and VMX
For runtime switching, the initial `Vm_session_component` on x86_64 used a
`Vm_page_table` implementation that for nested paging transparently
constructed a EPT for VMX or a HPT for SVN. While this worked, it came
at the cost of hidden complexity and side effects (in that the
implementation cached the virtualization type).

Split out the VMX an SVM specific parts (namely, the page table type) to
VMX and SVM specific implementations of the Vm_session_component
interface. To contain complexity, keep common functionality in a
`Vm_session_component` base class, which enables the continued sharing
of common method implementations across architectures.

Issue #5221